DE/data-pipelines/functions/schema-mapping-service/requirements.txt
# Azure Functions Core
azure-functions>=1.18.0

# Azure Storage (Data Lake Gen2 & Blob Storage)
azure-storage-file-datalake>=12.19.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Azure Key Vault (for secrets management)
azure-keyvault-secrets==4.7.0

# Encoding Detection
charset-normalizer>=3.3.2

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Internal Schema & Validation
pydantic>=2.5.0

# LLM Services (for later systems)
openai>=1.12.0
# azure-ai-openai>=1.0.0

# Database & Cache
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0

# Service Bus
azure-servicebus>=7.11.0
azure-mgmt-servicebus>=9.0.0

# Container Execution (for sandbox)
azure-mgmt-containerinstance>=10.1.0
azure-mgmt-resource>=23.0.0

# Utilities
python-dateutil>=2.8.2
